+++ src/main/java/com/example/trainingnew/reprository/OTPRepo.java
@@ -4,11 +4,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.trainingnew.model.OTPModel;
 
-public interface OTPRepo extends JpaRepository<OTPModel, Long>{
+public interface OTPRepo extends JpaRepository<OTPModel, Integer>{
 
-	OTPModel findOneByEmailAndOtp(String email, long otp);
+	OTPModel findOneByEmailAndOtp(String email, Integer otp);
 
-	void deleteByOtp(long otp);
+	void deleteByOtp(Integer otp);
 
 //	void deleteByOtp(long otp);
